[#20] refactor : QuestionCard를 components하위 디렉토리인 common폴더로 이동
